python/pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "contriwork-config-core"
description = "Cross-language static configuration loader — env + files + secret refs, schema-validated."
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
authors = [
  { name = "ContriWork", email = "maintainers@contriwork.dev" }
]
keywords = ["contriwork"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "Typing :: Typed",
]
dynamic = ["version"]
dependencies = [
  "pydantic>=2.6",
  "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/contriwork/contriwork-config-core"
Repository = "https://github.com/contriwork/contriwork-config-core"
Issues = "https://github.com/contriwork/contriwork-config-core/issues"
Changelog = "https://github.com/contriwork/contriwork-config-core/blob/main/CHANGELOG.md"

[dependency-groups]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.24",
  "ruff>=0.8",
  "mypy>=1.13",
  "pip-audit>=2.7",
  "types-PyYAML>=6.0",
]

[tool.hatch.version]
path = "../VERSION"
pattern = "^(?P<version>[0-9]+\\.[0-9]+\\.[0-9]+)"

[tool.hatch.build.targets.wheel]
packages = ["src/contriwork_config_core"]

[tool.hatch.build.targets.sdist]
include = [
  "src/",
  "tests/",
  "README.md",
  "pyproject.toml",
]

[tool.ruff]
line-length = 100
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E", "F", "W",        # pycodestyle + pyflakes
  "I",                  # isort
  "B",                  # bugbear
  "UP",                 # pyupgrade
  "S",                  # bandit-equivalent security checks
  "N",                  # pep8-naming
  "RUF",                # ruff-specific
  "ASYNC",              # async issues
  "SIM",                # simplify
]
ignore = ["S101"]       # allow assert in tests

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S", "N"]
# Exception class names mirror their stable CONTRACT.md error codes
# (VALIDATION_FAILED → ValidationFailed, etc.) rather than PEP 8's
# "SomethingError" convention. The code is the contract; the class name
# is local convention.
"src/contriwork_config_core/errors.py" = ["N818"]

[tool.mypy]
python_version = "3.13"
strict = true
packages = ["contriwork_config_core"]
mypy_path = "src"
warn_unreachable = true
warn_return_any = true
disallow_any_explicit = false
show_error_codes = true
pretty = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --strict-config"
markers = [
  "contract: cross-language contract tests driven by contract-tests/test_cases.json",
]
